backend/requirements.txt
flask==3.0.3
chromadb==0.5.5
requests==2.32.3
python-dotenv==1.0.1
sentence-transformers==3.0.1
pdfplumber==0.11.4